requirements.txt
language-tool-python==2.9.0
python-dotenv==1.0.0 